add: scene select windows, fix: YJ_1 decompress

+++ SwordStudio.NET/PalConfig.cs
@@ -3,16 +3,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-using BOOL	= System.Boolean;
-using CHAR	= System.Char;
-using BYTE	= System.Byte;
-using SHORT	= System.Int16;
-using WORD	= System.UInt16;
-using INT	= System.Int32;
-using UINT	= System.UInt32;
-using SDWORD	= System.Int64;
-using DWORD	= System.UInt64;
-using LPSTR	= System.String;
+using BOOL	    = System.Boolean;
+using CHAR	    = System.Char;
+using BYTE	    = System.Byte;
+using SHORT	    = System.Int16;
+using WORD	    = System.UInt16;
+using INT	    = System.Int32;
+using UINT	    = System.UInt32;
+using SDWORD    = System.Int32;
+using DWORD     = System.UInt32;
+using LPSTR	    = System.String;
 
 namespace PalConfig
 {
